Move ShowHidePrimitive into primitives directory
